requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
cffi==1.17.1
discord.py==2.5.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.3
propcache==0.3.0
pycparser==2.22
PyNaCl==1.5.0
typing_extensions==4.12.2
wavelink==3.4.1
yarl==1.18.3
yt-dlp==2025.2.19
